+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 * make RON support optional.
 * added support for sorting maps before serialization.
 * added settings support.
+* added support for overriding the snapshot path.